functional_testing/Grid_API_Testing/requirements.txt
termcolor
requests
paramiko
configparser
nose-testconfig
nose
git+https://github.com/gigforks/packet-python.git
